don't crash when a ParameterInput doesn't have a Parameter target, on_unbound callback for when a Parameter is unbound from a ParameterInput, uhh some other changes/fixes

+++ include/FrequencyParameter.h
@@ -26,8 +26,8 @@ class FrequencyParameter : public Parameter<TargetClass, DataType> {
             this->currentValue = value;
 
             if (this->debug) {
-                Serial.printf("FrequencyParameter#setParamValue(): calling setter func with frequency (%i) from input value ", freq);
-                Serial.println(value * 5.0);
+                Serial.printf("FrequencyParameter#setParamValue(): calling setter func with frequency (%i) from input value ", (uint32_t) freq);
+                Serial.println((value * 5.0));
             }
             //Serial.println();
